Extract pi-codex websocket transport
